Rename area.py to model.py
